requirements.txt
gevent
flask
requests
py-xmltv
xsdata[cli]
python-dotenv